Added naming for wallets, custom metrics doesn't work.

+++ puller/src/main/java/ru/hse/bot/web/dto/transfers/SolanaTransactionDataResponse.java
@@ -11,6 +11,6 @@ public record SolanaTransactionDataResponse(
         @NotNull String source,
         @Nullable String destination,
         @NotNull String token,
-        @NotNull Long amount,
+        @NotNull Double amount,
         @NotNull Long timestamp
 ) { }
